requirements.txt
fastapi
uvicorn
transformers
torch
numpy
pyyaml
huggingface-hub
llama-cpp-python